backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-dotenv==1.0.1
PyMuPDF==1.25.5
psycopg2-binary==2.9.9
sqlalchemy==2.0.28
pydantic==2.6.3
python-multipart==0.0.9
aiofiles==23.2.1 